with main menu

+++ Scripts/Main/build.sh
@@ -3,18 +3,23 @@
 # @Date:   06-03-2018 10:05
 # @Filename: config.sh
 # @Last modified by:   jorpilo
-# @Last modified time: 07-03-2018 13:17
+# @Last modified time: 06-04-2018 14:12
 
 # Check if we're root and re-execute if we're not.
-rootcheck () {
-    if [ $(id -u) != "0" ]
-    then
-        echo "Please run it as root: sudo ./config.sh"
-        exit 1
-    fi
-}
 
-rootcheck
+rootcheck () {
+        if [ $(id -u) != "0" ]
+        then
+            echo 1
+        else
+          echo 0
+        fi
+    }
+
+if [ $(rootchecks) -ne 0 ]; then
+    exit 1
+fi
+source config.conf
 if [ -d "built" ]; then
   echo "Warning: Cleaning built folder"
   rm -rf "built"
@@ -23,17 +28,16 @@ echo "Creating built folder"
 mkdir built
 cd built
 
-Mirror="http://softlibre.unizar.es/debian/"
-Locales="es_ES.UTF-8"
-Keyboard="es"
-UserName="live-user"
-Groups="audio,cdrom,dip,floppy,video,plugdev,netdev,powerdev,scanner,bluetooth,fuse"
+
+Options= $(echo config.conf)
+Mirror= $(echo main-repo.conf)
+
 Desktop=""
 lb config --binary-images iso-hybrid\
 --mirror-bootstrap $Mirror --mirror-binary $Mirror\
 --debian-installer live \
 --debootstrap-options "--include=apt-transport-https" \
---bootappend-live 'boot=live components username='"$UserName"'locales='"$Locales"'keyboard-layouts='"$Keyboard"'live-config.login live-config.user-default-groups='"$Groups"
+--bootappend-live 'boot=live components $Options'
 
 # Faltan desktops
 echo "task-spanish-desktop task-spanish" >> config/package-lists/desktop.list.chroot
@@ -41,5 +45,8 @@ echo "task-spanish-kde-desktop" >> config/package-lists/desktop.list.chroot
 cd ..
 # Configure packages
 
-../packages.sh
+packages.sh
 echo "Done"
+
+cd built
+lb build
